Performs thin-plate spline downscaling

The function Tpsdownscale is a thin plate spline model, typically with elevation as a covariate to downscale data.

Usage

Tpsdownscale(r, dtmc, dtmf, method = "normal", fast = TRUE)

Arguments

r

a single layer SpatRast dataset to be downscaled.

dtmc

a coarse resolution SpatRast of elevations matching the resolution, coordinate reference system and extent of r.

dtmf

a fine-resolution SpatRast of elevations.

method

one of normal, log or logit (see details)

fast

optional logical indicating whether to use fields::fastTps() (faster but less accurate)

Value

a SpatRast of r downscaled to match dtmf.

Details

if method = "log" data are log-transformed prior to performing the downscale, and then back-transformed. Use this method if input and output data must always be non-negative. if method = "logit" data are logit-transformed prior to performing the downscale, and then back-transformed. Use this method if input and output data must always be bounded by 0 and 1. In both instances, the spacial case where input data are 0 or 1 is handled. If method = "normal" no transformation is applied.

Examples

 if (FALSE) { # \dontrun{
climdata<-read_climdata(mesoclim::ukcpinput)
dtmf<-terra::rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim'))
rain<-climdata$prec
try(rainf<-Tpsdownscale(rain, climdata$dtm, dtmf, method = "normal", fast = TRUE))
terra::plot(rain,main='Input rain')
terra::plot(rainf,main='Downscaled rain')
} # }