Applies bias correction model to climate variable

The function biascorrect_apply applies a set of bias corrections derived using bias_correct to a spatial climate dataset

Usage

biascorrect_apply(fut_mod, biasmods, rangelims = NA)

Arguments

fut_mod

a SpatRaster of modelled climate data for for e.g. a future period to which corrections are applied.

biasmods

an object of class biascorrectmodels derived using biascorrect().

Value

a SpatRaster matching the overlap of fut_mod and biasmods$rindex

Details

The biascorrect_apply function allows the same bias correction to be applied to multiple datasets, spanning e.g. several future time periods or climate scenarious. The spatial extent of fut_mod must at least partially overlap with that of area used in biasmods when running biascorrect as corrected values are only returned for the overlapping area

See also