Interpolates sea surface temperature data to coastal regions Resamples sst data to same extent, resolution and projection as aoi Interpolates from mean of adjacent sea cells

Interpolates sea surface temperature data to coastal regions Resamples sst data to same extent, resolution and projection as aoi Interpolates from mean of adjacent sea cells

Usage

.sea_to_coast(sst.r, aoi.r, ext_cells = 8)

Arguments

sst.r

sea surface temperature data

aoi.r

dtm of area of interest where sea = NA

Value

sst spatrasters of same projection, res and extent as aoi with matching coastline

Examples

sst.r<-terra::rast(system.file("extdata/sst/NWSClim_NWSPPE_r001i1p00000_2018_gridT.nc",package="mesoclim"),"SST")[[1]]
#> Error: [rast] filename is empty. Provide a valid filename
aoi.r<-terra::unwrap(mesoclim::ukcpinput$dtm)
terra::plot(.sea_to_coast(sst.r,aoi.r))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'plot': error in evaluating the argument 'x' in selecting a method for function 'project': object 'sst.r' not found