Coastal temperature effects

Coastal temperature effects

Usage

.tempcoastal(
  tc,
  sstf,
  u2,
  wdir,
  dtmf,
  dtmm,
  dtmc,
  ndir = 32,
  smooth = 5,
  correct = TRUE
)

Arguments

tc

downscaled temperature at same res as dtmf

sstf

downscaled sea surface temperature to dtmf resolution and extent - no NA and timeseries must match tc

u2

downscaled windspeed at temprature height

wdir

wind direction (coarse resolution) - same value for all of dtmf extent will be used

dtmf

fine dtm spatraster

dtmm

fine-scale dtm covering wider area than dtmf (but same resolution!!)

dtmc

coarse dtm spatraster matching resolution and extent of wdir

ndir

number of directions to calculate for wind exposure

smooth

number of cells to use for smoothing matrix

correct

corrects each timestep so that output mean temp of area matches input area mean temp

Value

Spatraster of temperature that includes coastal effect