Download 1km albedo data

The function download_globalbedo downloads monthly GlobAlbedo tiled albedo data from Jasmin available for the years 1998-2011. Whole of the UK covered by default tiles. For further information see: http://www.globalbedo.org/index.php

Usage

download_globalbedo(
  dir_out,
  tiles = c("h17v03", "h17v04", "h18v03"),
  months = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12"),
  years = seq(1998, 2011),
  url = "http://gws-access.jasmin.ac.uk"
)

Arguments

dir_out

directory to which to save data

tiles

vector of strings of the file name in the format 'h##v##'. Whole of the UK covered by default tiles.

year

numeric vector indicating the years for which data are required

month

string vector indicating the month for which data are required

Value

database detailing download details and success of all requested files

Details

No username or password currently required.