Converts era5 hourly data to daily to enable bias correction to be applied to ukcp data

Converts era5 hourly data to daily to enable bias correction to be applied to ukcp data

Usage

era5todaily(
  filein,
  pathout,
  landsea,
  elev,
  output = c("tifs", "singlenc", "none"),
  vars = c("tasmax", "tasmin", "sp", "psl", "hurs", "huss", "u10", "v10", "swrad",
    "difrad", "lwrad", "skyem", "pr")
)

Arguments

filein

filename (including path ) of era5 nc file with required variables

pathout

directory in which to save data

landsea

a landsea raster object of land fractions that must match the extent (or less) and crs of the era5 data era5 data will be cropped to lsm extent

elev

a spatRaster of surface elevations to correct surface pressure to sea level pressure

output

character of value tifs, singlenc or none - determines type of files (if any) written

vars

variables for which daily values calculated

Value

list of daily spatRasters of each variable requested. Also write .tiff or .nc files if requested The function saves individual terra SpatRaster objects to disk (one for each variable)

Examples

 if (FALSE) { # \dontrun{
vars<-c('psl','hurs','u10','v10','swrad','difrad','lwrad')
daily<-.era5todaily(filein=file.path(dir_era5,f),tempdir(),landsea=lsm,output='none',vars)
} # }