Derives hourly temperature from daily data

Generates a plausible vector or SpatRaster of hourly temperatures from daily data.

Usage

hourlytemp(tmn, tmx, tme = NA, lat = NA, long = NA, srte = 0.09)

Arguments

tmn

a vector or SpatRaster of daily minimum temperatures (deg C)

tmx

a vector or SpatRaster of daily maximum temperatures (deg C)

tme

POSIXlt object of dates. Only required if tmn and tmx are vectors.

lat

latitude of location (decimal degrees). Only required if tmn and tmx are vectors

srte

a parameter controlling speed of decay of night time temperatures (see details)

lon

longitude of location (decimal degrees). Only required if tmn and tmx are vectors

Details

Daytime temperatures are assumed to follow a sine curve with a peak a short while after solar noon. After dusk, the temperatures are assumed to decay exponentially reaching a minimum at dawn. The day in which tmx and tmn fall is assumed to match UTC days. The parameter stre controls the speed of decay of night time temperatures with time. A value of zero ensures values drop to minimum at dawn the following day, but trial and error indicates in most circumstances temperatures decay faster than this. The default value of 0.09 is an optimal value derived using ERA5 data for western Europe, but performs reasonably well globally

Examples

if (FALSE) {
climdata<- read_climdata(system.file('extdata/era5/input.rds',package='mesoclim'))
tc<-climdata$temp
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as vector =========================== #
# ========================================================================= #
# Derive daily tmx and tmn from hourly era5 data
tch <- apply(tc, 3, mean, na.rm = TRUE)
tcm <- matrix(tch, ncol = 24, byrow = TRUE)
tmn <- apply(tcm, 1, min)
tmx <- apply(tcm, 1, max)
tme <- as.POSIXlt(c(0:30) * 3600 * 24, origin = "2018-05-01", tz = "UTC")
# Use interpolation function
tcp <- hourlytemp(tmn, tmx, tme, 50, -5)
tmeh <- as.POSIXct(climdata$tme)
# Plot results to compare
plot(tch ~ tmeh, type="l", ylim = c(8, 18), xlab = "Date", ylab = "Temperature")
par(new = TRUE)
plot(tcp ~ tmeh, type="l", ylim = c(8, 18), col = "red", xlab = "", ylab = "")
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as SpatRaster ======================= #
# ========================================================================= #
dtmc <- climdata$dtm
# Derive tmx and tmn from hourly and convert to spatraster
tmn <- .rast(.hourtodayCpp(tc, "min"),dtmc)
tmx <- .rast(.hourtodayCpp(tc, "max"),dtmc)
# Use interpolation function
tp <- hourlytemp(tmn, tmx, tme)
# Plot results to compare
tc <- .rast(tc,dtmc)
par(mfrow=c(2,1))
plot(tc[[4]]-tp[[4]],main='Difference (tc-tp) for coldest tc hour')
plot(tc[[665]]-tp[[665]],main='Difference (tc-tp) for hottest tc hour')
}