Daily to hourly relative humidity

Daily to hourly relative humidity

Usage

hum_dailytohourly(
  relhum,
  tasmin,
  tasmax,
  temph,
  psl,
  presh,
  tme,
  relmin = 2,
  adjust = TRUE
)

Arguments

relhum

an array of daily relative humidity values (%)

tasmin

an array of daily minimum temperature values (deg C)

tasmax

an array of daily maximum temperature values (deg C)

temph

an array of hourly temperature values (deg C)

psl

an array of daily surface-level pressure values(kPa)

presh

an aray of hourly surface-level pressure values (kPa)

tme

POSIXlt object of dates corresponding to radsw

relmin

minimum possible relative humidity value

adjust

optional logical which if TRUE ensures that, after interpolation, returned hourly values, when averaged to daily, match the input

Value

an array of hourly relative humidity values (%)

Details

#' Owing to the strong dependence of relative humidity on temperature, the resulting diurnal patterns, prior to interpolation, relative humidity is first converted to specific humidity using tasmin, tasmax and psl. After interpolation, the data are back-converted to relative humidity, using temph and presh. ~~ Function should spline interpolate vapour pressure and use temperature cycle Derives an array of hourly relative humidity values from daily values.

Examples

if (FALSE) {
climdata<- read_climdata(system.file('extdata/preprepdata/ukcp18rcm.Rds',package='mesoclim'))
# Get hourly variables required for rel hum downscaling
dtmc<-climdata$dtm
temph<-as.array(hourlytemp(mesoclim:::.rast(climdata$tmin,dtmc), mesoclim:::.rast(climdata$tmax,dtmc), climdata$tme, 50, -2.5))
presh<-pres_dailytohourly(climdata$pres,climdata$tme)
# Get sea level pressure - MAKE THIS A FUNCTION!!
psl_r<-mesoclim:::.rast(climdata$pres,dtmc) / (((293-0.0065*ifel(is.na(dtmc),0,dtmc))/293)^5.26)
psl<-mesoclim:::.rta(psl_r,length(climdata$tme))
humh<-hum_dailytohourly(climdata$relhum, climdata$tmin, climdata$tmax, temph, psl, presh, climdata$tme)
plot_q_layers(mesoclim:::.rast(humh,dtmc))
}