Daily to hourly downward longwave radiation

Derives an array or spatraster of hourly effective downward longwave radiation values.

Usage

lw_dailytohourly(lw, hrtemps, hrrh, hrpres, tme = NA, adjust = TRUE)

Arguments

lw

an array of daily mean sky-emissivity values values (0-1)

hrtemps

hourly temperature data for corresponding period as lw

hrrh

hourly relative humidity data for corresponding period as lw

hrpres

hourly atmospheric pressure data for corresponding period as lw

tme

POSIXlt object of dates corresponding to radsw

adjust

optional logical which if TRUE ensures that, after interpolation, returned hourly values, when averaged to daily, match the input

dtm

digital elevation spatraster

Value

an array of hourly downward long wave radiation

Details

Effective sky emissvity can be used to calaculate downward longwave radiation (Lwd). The formula is Lwd = skyem * Lwu where Lwu is upward longwave radiation given by Lwu=0.975.6710**-8*(tc+273.15). Here tc is average surface temperature (deg C)) but an adequate approximation is derived if substituted by air temperature. TO DO - Option of providing cloud cover data to correct sky emissivity values??

Examples

# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as vector =========================== #
# ========================================================================= #
climdaily<-read_climdata(ukcpinput)
lw<-unlist(global(climdaily$lwrad,mean))
lw<-unlist(global(climdaily$lwrad,mean))
lw<-unlist(global(climdaily$lwrad,mean))
lw<-unlist(global(climdaily$lwrad,mean))
tme<-as.POSIXlt(terra::time(climdaily$swrad))
r<-c(50,-5)
hrsw <- lwrad_dailytohourly(radsw,tme,r)
#> Error in lwrad_dailytohourly(radsw, tme, r): could not find function "lwrad_dailytohourly"
hr_tme<-as.POSIXlt(unlist(lapply(tme,FUN=function(x) x+(60*60*c(0:23)) )))
# Plot results to compare
matplot(x=as.numeric(hr_tme),y=hrsw, type = "l", lty = 1)
#> Error: object 'hrsw' not found
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as SpatRaster ======================= #
# ========================================================================= #
hrtemps<-temp_dailytohourly(climdaily$tmin, climdaily$tmax, srte = 0.09)
hrpres<-pres_dailytohourly(climdaily$pres)
hrrh <- hum_dailytohourly(climdaily$relhum, climdaily$tmin, climdaily$tmax,hr_temp,climdaily$pres, hr_pres,relmin = 10)
#> Error: object 'hr_temp' not found
lwdhr<-lw_dailytohourly(lw=climdaily$lwrad, hrtemps=hrtemps, hrrh=hrrh, hrpres=hrpres, adjust = TRUE)
#> Error: object 'hrrh' not found
# Plot results for one cell
cell_lw<-t(extract(lwdhr,matrix(c(175000,40000),ncol=2)))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 't': error in evaluating the argument 'x' in selecting a method for function 'extract': object 'lwdhr' not found
matplot(as_datetime(time(lwdhr)),cell_lw, type = "l", lty = 1)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'as_datetime': error in evaluating the argument 'x' in selecting a method for function 'time': object 'lwdhr' not found
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as 3D array ======================= #
# ========================================================================= #
hrlw <- lw_dailytohourly(lw=.is(climdaily$lwrad), hrtemps=.is(hrtemps), hrrh=.is(hrrh), hrpres=.is(hrpres), tme=climdaily$tme,adjust = TRUE)
#> Error in .is(climdaily$lwrad): could not find function ".is"
cell_lw<-hrlw[3,4,]
#> Error: object 'hrlw' not found
hr_tme<-as.POSIXlt(unlist(lapply(climdaily$tme,FUN=function(x) x+(60*60*c(0:23)) )))
matplot(x=as.numeric(hr_tme),y=cell_lw, type = "l", lty = 1)
#> Error: object 'cell_lw' not found
# ========================================================================= #
# ~~~~~~~~~~~~~~~~~~~~ input provided as vector =========================== #
# ========================================================================= #
lw<-unlist(global(climdaily$lwrad,mean))
hrt<-unlist(global(hrtemps,mean))
hrp<-unlist(global(hrpres,mean))
hrh<-unlist(global(hrrh,mean))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'global': object 'hrrh' not found
tme<-as.POSIXlt(terra::time(climdaily$swrad))
hrlw <- lw_dailytohourly(lw=lw, hrtemps=hrt, hrrh=hrh, hrpres=hrp, tme=tme, adjust = TRUE)
#> Error: object 'hrh' not found
hr_tme<-as.POSIXlt(unlist(lapply(tme,FUN=function(x) x+(60*60*c(0:23)) )))
# Plot results to compare
matplot(x=as.numeric(hr_tme),y=hrlw, type = "l", lty = 1)
#> Error: object 'hrlw' not found