Applies bias correction to precipitation data

The function precipcorrect applies bias correction to spatial climate dataset

Usage

precipcorrect(
  hist_obs,
  hist_mod,
  fut_mod = NA,
  mod_out = FALSE,
  rangelim = NA,
  prec_thold = 0.01
)

Arguments

hist_obs

a stacked SpatRast of precipitation data.

hist_mod

a stacked SpatRast of modelled precipitation data for for the same period and spatial area as hist_obs. Reprojected and resampled to match hist_obs if necessary.

fut_mod

a stacked SpatRast of modelled precipitation data for for e.g. a future period to which corrections are applied. eprojected and resampled to match hist_obs if necessary.

mod_out

optional logical indicating whether to return model correction coefficients, which can be applied to datasets over multiple time periods using precipcorrect_apply().

rangelim

fractional amount above the maximum values of precipitation in hist_obs in any given cell by which values in the returned dataset can attain. I.e. if set to 2, values in any given grid cell of the returned dataset cannot exceed twice the maximum amount in the corresponding grid cell in hist_obs.If not supplied (the default), returned values are unbounded.

prec_thold

numeric value (mm per day) below which daily precipitation is taken to be zero.

Value

If mod_out = FALSE (the default), a SpatRaster of bias corrected precipitation. If mod_out = TRUE a list containing two wrapped SpatRasters representing the amount by which to adjust precipitation totals and precipitation day fractions.

See also