Applies bias correction model to precipitation data

The function precipcorrect_apply applies a set of bias corrections derived using precip_correct to a spatial climate dataset

Usage

precipcorrect_apply(fut_mod, biasmods, prec_thold = 0.01)

Arguments

fut_mod

a stacked SpatRast of modelled precipitation data for for e.g. a future period to which corrections are applied. reprojected and resampled to match data held in mod_out if necessary.

biasmods

list containing two wrapped SpatRasters derived using precipcorrect() representing the amount by which to adjust rainfall totals and rainfall days, which is used as a basis for adjusting individual rainfall events.

prec_thold

numeric value (mm per day) below which precipitation is taken to be zero

Value

a SpatRaster matching fut_mod with climate data bias corrected.

Details

The precipcorrect_apply function allows the same bias correction to be applied to multiple datasets, spanning e.g. several future time periods or climate scenarious. The spatial extent of fut_mod must at least partially overlap with that of hist_obs when running precipcorrect as corrected values are only returned for the overlapping area

See also