Downscale relative humidity accounting for downscaled temperature

The function relhumdownscale is used to spatially downscale relative

Usage

relhumdownscale(rh, tcc, tcf, dtmc, rhmin = 20)

Arguments

rh

a coarse-resolution array of relative humidities (percentage)

tcc

a coarse-resolution array of temperatures (deg C)

tcf

a fine-resolution SpatRast of temperatures (deg C) as returned by temphrly_downscale() or tempdaily_downscale()

dtmc

a coarse-resolution SpatRast of elevations matching the resolution, extent and coordinate reference system of rh and tc.

rhmin

optional single numeric value indicating the minimum realistically attainable relative humidity within the study region (percentage).

Value

a stacked SpatRast of relative humidities (percentage) matching the resolution, coordinate reference system and extent of tcf.

See also

Examples

dtmf<-terra::rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim'))
climdata<-read_climdata(mesoclim::ukcpinput)
tmean_c<-.hourtoday(temp_dailytohourly(climdata$tmin, climdata$tmax, climdata$tme),mean)
#> Error in .hourtoday(temp_dailytohourly(climdata$tmin, climdata$tmax, climdata$tme),     mean): could not find function ".hourtoday"
tcf<-tempdaily_downscale(climdata,sst=NA,dtmf=dtmf,dtmm=NA,cad=FALSE,coastal=FALSE)$tmean
rhf<-relhumdownscale(climdata$relhum,tcc=tmean_c,tcf=tcf,dtmc=climdata$dtm,rhmin=20)
#> Error: object 'tmean_c' not found
panel(c(rhf[[1]],rhf[[5]],rhf[[10]],rhf[[20]]),main=paste0("Rel humidity ",c(1,5,10,20),"/05/2018"))
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'panel': object 'rhf' not found