Daily to hourly downward longwave radiation

Derives an array of hourly effective sky-emissivity values.

Usage

skyem_dailytohourly(skyem, tme, adjust = TRUE)

Arguments

skyem

an array of daily mean sky-emissivity values values (0-1)

tme

POSIXlt object of dates corresponding to radsw

adjust

optional logical which if TRUE ensures that, after interpolation, returned hourly values, when averaged to daily, match the input

Value

an array of hourly sky-emissivity values values (0-1)

Details

NB - more consistent to code this as downward longwave, but will essentially do the calaculations in the function below, but with temperature as an additional Effective sky emissvity can be used to calaculate downward longwave radiation (Lwd). The formula is Lwd = skyem * Lwu where Lwu is upward longwave radiation given by Lwu=0.975.6710**-8*(tc+273.15). Here tc is average surface temperature (deg C)) but an adequate approximation is derived if subtited by air temperature.

Examples

climdata<- read_climdata(system.file('data/ukcpinput.rds',package='mesoclim'))
#> Error in read_climdata(system.file("data/ukcpinput.rds", package = "mesoclim")): Filepath provided does NOT exist!!
lwrad<-skyem_dailytohourly()
#> Error in skyem_dailytohourly(): argument "skyem" is missing, with no default