Downscale shortwave radiation

Downscales an array of shortwave radiation data with option to simulate cloud patchiness and calculate diffuse radiation

Usage

swdownscale(
  swrad,
  tme,
  dtmf,
  dtmc,
  patchsim = FALSE,
  nsim = dim(swrad)[3],
  terrainshade = FALSE
)

Arguments

swrad

an array of shortwave radiation (W/m^2).

tme

POSIXlt object of times corresponding to radiation values in swrad.

dtmf

a fine-resolution SpatRast of elevations. Temperatures down-scaled to resolution of dtmf.

dtmc

SpatRast of elevations matching the resolution, extend and coordinate reference system of of swrad.

patchsim

optional logical indicating whether to simulate cloud cover patchiness during downscaling. More realistically captures variation, but slower.

nsim

optionally the number of independent cloud cover patchiness simulations to perform. Outputs are temporally interpolated (see details).

terrainshade

optional logical indicating whether to adjust incoming radiation for terrain shading. If TRUE, total downward shortwave radiation is partioned into its direct and diffuse components, the latter adjusted by sky view and the former set to zero if the sun is below the horizon, which is computed explicitly.

Value

if terrainshade = FALSE a multi-layer SpatRast of shortwave radiation (W/m^2) matching the resolution of dtmf. If if terrainshade = TRUE a list of wrapped multi-layer SpatRasts of (1) shortwave radiation (W/m^2) and (2) the diffuse fraction (range 0-1).

Details

radiation is downscaled by computing average fraction of clear-sky radiation and then adjusting this variable to account for elevation using an emprical adjustment calibrated against 0.05 degree data for the UK. If patchsim is set to TRUE cloud cover patchiness is simulated the gstats package. The parameter nsim determines the number of independent simulations and hence the time intervals at which these simulations are performed. Simulated anomalies due to local cloud patchiness are then interpolated temporally between these periods. This ensures that, over shorter increments of say and hour, the location clouds within each hour retain a degree of inter-dependence more realistically simulating the trajectory of cloud as they move across the landscape.

Examples

climdata<- read_climdata(system.file('extdata/preprepdata/ukcp18rcm.Rds',package='mesoclim'))
swradf<-swdownscale(climdata$swrad, climdata$tme,  rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim')), ukcpinput$dtm)
#> Warning: is.na() applied to non-(list or vector) of type 'S4'
#> Error in (function (classes, fdef, mtable) {    methods <- .findInheritedMethods(classes, fdef, mtable)    if (length(methods) == 1L)         return(methods[[1L]])    else if (length(methods) == 0L) {        cnames <- paste0("\"", vapply(classes, as.character,             ""), "\"", collapse = ", ")        stop(gettextf("unable to find an inherited method for function %s for signature %s",             sQuote(fdef@generic), sQuote(cnames)), domain = NA)    }    else stop("Internal error in finding inherited methods; didn't return a unique method",         domain = NA)})(list("logical"), new("standardGeneric", .Data = function (test,     ...) standardGeneric("ifel"), generic = structure("ifel", package = "terra"),     package = "terra", group = list(), valueClass = character(0),     signature = "test", default = NULL, skeleton = (function (test,         ...)     stop(gettextf("invalid call in method dispatch to '%s' (no default method)",         "ifel"), domain = NA))(test, ...)), <environment>): unable to find an inherited method for function ‘ifel’ for signature ‘"logical"’
plot_q_layers(swradf)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'global': object 'swradf' not found