convert UKCP18 ncdf4 files to format required for model

Converts UKCP18 global or regional land data in the form of netCDF4 file (as returned by ukcp18_downscale() to the correct data form required for subsequent modelling.

Usage

ukcp18toclimarray(
  dir_ukcp,
  dtm,
  startdate,
  enddate,
  collection = c("land-gcm", "land-rcm"),
  domain = c("uk", "eur", "global"),
  member = c("01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", "12",
    "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26",
    "27", "28"),
  wsalbedo = 0.19,
  bsalbedo = 0.22,
  ukcp_vars = c("clt", "hurs", "pr", "psl", "rls", "rss", "tasmax", "tasmin", "uas",
    "vas"),
  ukcp_units = c("%", "%", "mm/day", "hPa", "watt/m^2", "watt/m^2", "degC", "degC",
    "m/s", "m/s"),
  output_units = c("%", "%", "mm/day", "kPa", "watt/m^2", "watt/m^2", "degC", "degC",
    "m/s", "m/s"),
  toArrays = TRUE,
  sampleplot = FALSE
)

Arguments

dir_ukcp

directory holding ALL ncdf files of UKCP18 data required to extract the requested variables and timeseries.

dtm

SpatRaster object of elevations covering the geographical extent required (see details)

startdate

POSIXlt class defining start date of required timeseries

enddate

POSIXlt class defining end date of required timeseries

collection

text string defining UKCP18 collection, either 'land-gcm' or 'land-rcm'

domain

text string defining UKCP18 domain, either 'uk' or 'eur'(land-rcm collection only) or 'global'

member

string defining the climate model member to be used for the timeseries. Available members vary between UKCP18 collections.

ukcp_vars

UKCP18 variable names to be extracted DO NOT CHANGE

ukcp_units

units of the UKCP18 variables extracted DO NOT CHANGE

output_units

units required for output DO NOT CHANGE CHECK CORRECT

toArrays

logical determining if climate data returned as list of arrays. If FALSE returns list of Spatrasts.

sampleplot

if TRUE plots examples of interpolated dates when converting from 360 to 366 day years

Value

a list of the following:

dtm

Digital elevation of downscaled area in metres (as Spatraster)

tme

POSIXlt object of times corresponding to climate observations

windheight_m

Height of windspeed data in metres above ground (as numeric)

tempheight_m

Height of temperature data in metres above ground (as numeric)

temp

Temperature (deg C)

relhum

Relative humidity (Percentage)

pres

Sea-level atmospheric pressure (kPa)

swrad

Total downward shortwave radiation (W/m^2)

difrad

Downward diffuse radiation (W / m^2)

lwrad

Total downward longwave radiation (W/m^2)

windspeed

At windheight_m above ground` (m/s)

winddir

Wind direction (decimal degrees)

prec

Precipitation (mm)

Details

The function converts 360 day UKCP18 to actual calendar data by reassignment and linear interpolation of missing data. Returned climate data will be at the extent, corrdinate reference system of the UKCP18 data requested, which can vary between domains and collections. The extent will be provided by the dtm which, if necessary, will be re-projected and resampled to the CRS and resolution of UKCP18 data requested. For regional UKCP18 data, it is recommended that 'dtm' is derived from the original orography data available for download.

Examples

dir_ukcp<-system.file('extdata/ukcp18rcm',package='mesoclim')
dtm<-terra::rast(system.file('extdata/ukcp18rcm/orog_land-rcm_uk_12km_osgb.nc',package='mesoclim'))
dtm<-crop(dtm,c(28077.86, 339436.5, -57947.13, 204298.7))
#> Error in crop(dtm, c(28077.86, 339436.5, -57947.13, 204298.7)): could not find function "crop"
ukcpinput<-ukcp18toclimarray(dir_ukcp,dtm,as.POSIXlt('2018/05/01'),as.POSIXlt('2018/05/31'),collection='land-rcm',domain='uk',member='01')
#> Loading clt_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading hurs_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading pr_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading psl_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading rls_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading rss_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading tasmax_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading tasmin_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading uas_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Loading vas_rcp85_land-rcm_uk_12km_01_day_20101201-20201130.nc
#> Using constant land and sea albedo values - assuming NA values in dtmc are sea!!