derive wind terrain adjustment coefficient

The function windelev is used to spatially downscale wind, and adjusts wind speed for elevation and applies a terrain shelter coefficient for a specified wind direction.

Usage

windelev(dtmf, dtmm, dtmc, wdir, uz = 2)

Arguments

dtmf

a high-resolution SpatRast of elevations

dtmm

a medium-resolution SpatRast of elevations covering a larger area than dtmf (see details)

dtmc

a coarse-resolution SpatRast of elevations usually matching the resolution of climate data used for downscaling (see details)

wdir

wind direction (from, decimal degrees).

uz

height above ground (m) of wind speed measurement

Value

a SpatRast of wind adjustment coefficients matching the resoltuion, coordinate reference system and extent of dtmf.

Details

Elevation effects are derived by sampling the dtms at intervals in an upwind direction, determining the elevation difference form each focal cell and performing a standard wind-height adjustment. Terrain sheltering is computed from horizon angles following the method detailed in Maclean et al (2019) Methods Ecol Evol 10: 280-290. By supplying three dtms, the algorithm is able to account for elevation differences outside the boundaries of dtmf. The area covered by dtmm was extend at least one dtmc grid cell beyond dtmf. Elevations must be in metres. The coordinate reference system of dtmf must be such that x and y are also in metres. dtmm and dtmc are reprojected to match the coordinate reference system of dtmf.

See also

Examples

dtmf<-rast(system.file('extdata/dtms/dtmf.tif',package='mesoclim'))
#> Error in rast(system.file("extdata/dtms/dtmf.tif", package = "mesoclim")): could not find function "rast"
dtmm<-rast(system.file('extdata/dtms/dtmm.tif',package='mesoclim'))
#> Error in rast(system.file("extdata/dtms/dtmm.tif", package = "mesoclim")): could not find function "rast"
climdata<-read_climdata(system.file('data/ukcpinput.rds',package='mesoclim'))
#> Error in read_climdata(system.file("data/ukcpinput.rds", package = "mesoclim")): Filepath provided does NOT exist!!
wc <- windelev(dtmf, dtmm, climdata$dtm, wdir = 270)
#> Error in h(simpleError(msg, call)): error in evaluating the argument 'x' in selecting a method for function 'crs': object 'dtmm' not found
plot(wc)
#> Error in eval(expr, envir, enclos): object 'wc' not found